+++ includes/class-newspack-newsletters.php
@@ -394,6 +394,22 @@ final class Newspack_Newsletters {
 		);
 		\register_meta(
 			'post',
+			'text_color',
+			[
+				'object_subtype' => self::NEWSPACK_NEWSLETTERS_CPT,
+				'show_in_rest'   => [
+					'schema' => [
+						'context' => [ 'edit' ],
+					],
+				],
+				'type'           => 'string',
+				'single'         => true,
+				'auth_callback'  => '__return_true',
+				'default'        => '',
+			]
+		);
+		\register_meta(
+			'post',
 			'preview_text',
 			[
 				'object_subtype' => self::NEWSPACK_NEWSLETTERS_CPT,
@@ -516,7 +532,7 @@ final class Newspack_Newsletters {
 			'show_in_rest'     => true,
 			'supports'         => $supports,
 			'taxonomies'       => [ 'category', 'post_tag' ],
-			'menu_icon'        => 'data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGhlaWdodD0iMjQiIHZpZXdCb3g9IjAgMCAyNCAyNCIgd2lkdGg9IjI0Ij48cGF0aCB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGQ9Ik0yMS45OSA4YzAtLjcyLS4zNy0xLjM1LS45NC0xLjdMMTIgMSAyLjk1IDYuM0MyLjM4IDYuNjUgMiA3LjI4IDIgOHYxMGMwIDEuMS45IDIgMiAyaDE2YzEuMSAwIDItLjkgMi0ybC0uMDEtMTB6TTEyIDEzTDMuNzQgNy44NCAxMiAzbDguMjYgNC44NEwxMiAxM3oiIGZpbGw9IiNhMGE1YWEiLz48L3N2Zz4K',
+			'menu_icon'        => 'data:image/svg+xml;base64,' . base64_encode( '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" aria-hidden="true" focusable="false" fill="none"><path fill-rule="evenodd" clip-rule="evenodd" d="M3 7c0-1.1.9-2 2-2h14a2 2 0 0 1 2 2v10a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V7Zm2-.5h14c.3 0 .5.2.5.5v1L12 13.5 4.5 7.9V7c0-.3.2-.5.5-.5Zm-.5 3.3V17c0 .3.2.5.5.5h14c.3 0 .5-.2.5-.5V9.8L12 15.4 4.5 9.8Z"></path></svg>' ),
 		];
 		\register_post_type( self::NEWSPACK_NEWSLETTERS_CPT, $cpt_args );
 	}
@@ -973,11 +989,13 @@ final class Newspack_Newsletters {
 			$font_header      = get_post_meta( $post->ID, 'font_header', true );
 			$font_body        = get_post_meta( $post->ID, 'font_body', true );
 			$background_color = get_post_meta( $post->ID, 'background_color', true );
+			$text_color       = get_post_meta( $post->ID, 'text_color', true );
 			?>
 				<style>
 					.main-content {
 						background-color: <?php echo esc_attr( $background_color ); ?>;
 						font-family: <?php echo esc_attr( $font_body ); ?>;
+						color: <?php echo esc_attr( $text_color ); ?>;
 					}
 					.main-content h1,
 					.main-content h2,